Fix a back-link bug
A bug, which creates the empty-back-link when the page.back_page property is not declared, has been found.
It seems like the 'find' filter of the liquid does not work properly.
Now, back-link.html uses both 'where' and 'first' filters to create the proper back link, only when the page.back_page is declared in the current page.